backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain
langchain-community
langchain-qdrant
langchain-openai
langgraph

qdrant-client
sqlite-utils

streamlit
requests

mixedbread
sqlalchemy

pypdf
python-multipart